# backend/app/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: DataDabble API
  description: API for creating custom databases with flexible schemas, managing teams, billing, and visualizations.
  version: "1.0.0"
  contact:
    name: DataDabble Support
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /oauth2/authorize:
    get:
      tags: [OAuth 2.0]
      summary: Authorization endpoint - get consent screen details
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: scope
          in: query
          schema:
            type: string
            description: Space-separated scopes
        - name: state
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          schema:
            type: string
            description: PKCE code challenge
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256, plain]
            default: S256
      responses:
        "200":
          description: Authorization details for consent screen
          content:
            application/json:
              schema:
                type: object
                properties:
                  client:
                    type: object
                    properties:
                      client_id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                  scopes:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        description:
                          type: string
                  redirect_uri:
                    type: string
                  state:
                    type: string
        "400":
          description: Invalid request, client, or scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [OAuth 2.0]
      summary: Process user consent and issue authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, redirect_uri, approved]
              properties:
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                scopes:
                  type: array
                  items:
                    type: string
                state:
                  type: string
                approved:
                  type: boolean
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                  enum: [S256, plain]
      responses:
        "200":
          description: Authorization code or access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect_uri:
                    type: string
                  code:
                    type: string
                    description: Present only if approved
                  state:
                    type: string
                  error:
                    type: string
                    description: "access_denied if not approved"
        "400":
          description: Invalid request or client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /oauth2/token:
    post:
      tags: [OAuth 2.0]
      summary: Exchange authorization code or refresh token for access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type, client_id]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token]
                code:
                  type: string
                  description: Required for authorization_code grant
                client_id:
                  type: string
                client_secret:
                  type: string
                  description: Required unless using PKCE
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
                  description: PKCE verifier
                refresh_token:
                  type: string
                  description: Required for refresh_token grant
      responses:
        "200":
          description: Token response
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                    example: 3600
                  refresh_token:
                    type: string
                  scope:
                    type: string
        "400":
          description: Invalid grant or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
  /oauth2/revoke:
    post:
      tags: [OAuth 2.0]
      summary: Revoke an access or refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Token revoked (always returns 200 per RFC 7009)
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
  /oauth2/userinfo:
    get:
      tags: [OAuth 2.0]
      summary: Get current user info (standard OpenID Connect userinfo)
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub:
                    type: string
                    description: User ID
                  email:
                    type: string
                    format: email
                  name:
                    type: string
                  given_name:
                    type: string
                  family_name:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

# ════════════════════════════════════════════════
# Components
# ════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/v1/oauth2/authorize
          tokenUrl: /api/v1/oauth2/token
          scopes:
            "read:user": Read user profile
            "read:databases": Read databases
            "write:databases": Create, update, and delete databases
            "read:fields": Read fields
            "write:fields": Create, update, and delete fields
            "read:entries": Read entries
            "write:entries": Create, update, and delete entries
            "read:visualizations": Read visualizations
            "write:visualizations": Create, update, and delete visualizations
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        messages:
          type: object
          description: Field-level validation errors
          additionalProperties: true
    OAuthError:
      type: object
      properties:
        error:
          type: string
          example: invalid_request
        error_description:
          type: string
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Missing or invalid token
...